chore: Standardized tests
