+++ packages/ast/template/tests/helpers/index.ts
@@ -1,1 +1,33 @@
-export * from './handlebars.js';
+import { AST } from '../../src/index.js';
+
+export function transform(file: string): string {
+  const traverse = AST.traverse();
+
+  const ast = traverse(file, {
+    AttrNode(node) {
+      if (node.name !== 'local-class') {
+        return;
+      }
+
+      if (node.value.type === 'TextNode') {
+        node.name = 'class';
+
+        const attributeValue = node.value.chars.trim();
+
+        node.value = AST.builders.mustache(
+          AST.builders.path(`this.styles.${attributeValue}`),
+        );
+      }
+    },
+  });
+
+  return AST.print(ast);
+}
+
+export function traverse(file: string): string {
+  const traverse = AST.traverse();
+
+  const ast = traverse(file);
+
+  return AST.print(ast);
+}
