+++ packages/ast/template/tests/index/transform-file-is-empty.test.ts
@@ -1,11 +1,11 @@
 import { assert, test } from '@codemod-utils/tests';
 
-import { transformHandlebars } from '../helpers/index.js';
+import { transform } from '../helpers/index.js';
 
 test('index > transform (file is empty)', function () {
   const oldFile = '';
 
-  const newFile = transformHandlebars(oldFile);
+  const newFile = transform(oldFile);
 
   assert.strictEqual(newFile, '');
 });
