+++ packages/ast/template/tests/index/traverse-file-is-empty.test.ts
@@ -1,11 +1,11 @@
 import { assert, test } from '@codemod-utils/tests';
 
-import { traverseHandlebars } from '../helpers/index.js';
+import { traverse } from '../helpers/index.js';
 
 test('index > traverse (file is empty)', function () {
   const oldFile = '';
 
-  const newFile = traverseHandlebars(oldFile);
+  const newFile = traverse(oldFile);
 
   assert.strictEqual(newFile, '');
 });
